# backend/tests/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Customer": {
        "properties": {
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          }
        },
        "required": [
          "customer_id",
          "first_name",
          "last_name"
        ],
        "title": "Customer",
        "type": "object"
      },
...
      "Dispute": {
        "description": "A case. status: open, auto_resolved (the policy resolved it), escalated (waiting for a specialist), in_progress (a specialist took it) or closed.",
        "properties": {
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "evidence": {
            "additionalProperties": true,
            "description": "The transaction as it was, plus resolution (auto_resolved) or handoff (escalated).",
            "title": "Evidence",
            "type": "object"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null for an escalation that could not be tied to a transaction.",
            "title": "Transaction Id"
          }
        },
        "required": [
          "case_id",
          "customer_id",
          "reason_code",
          "summary",
          "status",
          "evidence",
          "created_at"
        ],
        "title": "Dispute",
        "type": "object"
      },
      "DisputeCreate": {
        "properties": {
          "reason_code": {
            "description": "e.g. unrecognized_charge, duplicate, failed, other",
            "minLength": 1,
            "title": "Reason Code",
            "type": "string"
          },
          "summary": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "transaction_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          }
        },
        "required": [
          "reason_code",
          "summary"
        ],
        "title": "DisputeCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Transaction": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "amount_usd_effective": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Usd Effective"
          },
          "case_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when the transaction already has a dispute case.",
            "title": "Case Id"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "dispute_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispute Status"
          },
          "merchant_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant Category"
          },
          "merchant_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant Name"
          },
          "process_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Process Date"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "response_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Code"
          },
          "response_meaning": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Standard (ISO 8583) meaning of response_code, as {es, pt}. An assumption: the organizer does not define the codes. Null when the code is empty or unknown.",
            "title": "Response Meaning"
          },
          "transaction_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Category"
          },
          "transaction_city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction City"
          },
          "transaction_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Country"
          },
          "transaction_date": {
            "format": "date-time",
            "title": "Transaction Date",
            "type": "string"
          },
          "transaction_id": {
            "title": "Transaction Id",
            "type": "string"
          },
          "transaction_status": {
            "title": "Transaction Status",
            "type": "string"
          },
          "transaction_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Type"
          }
        },
        "required": [
          "transaction_id",
          "transaction_date",
          "amount",
          "currency",
          "transaction_status"
        ],
        "title": "Transaction",
        "type": "object"
      },
      "TransactionPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass it as ?cursor= to get the next page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "TransactionPage",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nMock banking service (tool layer) for the Factored AI & Data Hackathon 2026\nprototype. Sandbox limitations, stated explicitly:\n\n- Login is a test session: `POST /session` checks `customer_id` +\n  `document_number` against `gold.customers` and issues a short-lived JWT\n  (HS256). No real identity provider is behind it.\n- Admin login (`POST /admin/session`) checks credentials against `ADMIN_USERS`\n  (bcrypt hashes, in-memory rate limit, single-process sandbox limitation).\n  Admins work an inbox of escalated cases with audited transitions\n  (claim -> in_progress -> closed); they never set `auto_resolved`.\n- `GET /meta/demo-scenarios` is a public demo aid: it surfaces only the fields\n  the test login itself asks for (customer_id, document_number, name), never\n  fraud columns.\n- Reads come from the read-only `gold.*` tables published by the dbt pipeline;\n  operational tables (`app.*`) are built by the versioned SQL migrations in\n  `app/migrations`, applied at startup under an advisory lock.\n- `/v1` is what the web app uses (user name and password login, products,\n  paginated history, summary). The routes at the root are the agent's.\n- No money is ever moved: disputes only explain, document and escalate.\n- Fraud ground truth (`is_fraud`/`fraud_score`) is not part of `gold.transactions`\n  and is never exposed here.\n",
    "title": "mock-banking-service",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/disputes": {
      "get": {
        "description": "The customer's own cases, newest first.",
        "operationId": "list_disputes_disputes_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(open|auto_resolved|escalated|in_progress)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Dispute"
                  },
                  "title": "Response List Disputes Disputes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Disputes",
        "tags": [
          "disputes"
        ]
      },
      "post": {
        "description": "Open the case. One per (customer, transaction): reporting the same transaction again returns that case.",
        "operationId": "create_dispute_disputes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
...
    "/me": {
      "get": {
        "operationId": "get_me_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "customers"
        ]
      }
    },
    "/me/disputes": {
      "get": {
        "operationId": "list_my_disputes_me_disputes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List My Disputes Me Disputes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Disputes",
        "tags": [
          "customers"
        ]
      }
    },
    "/me/transactions": {
      "get": {
        "operationId": "list_transactions_me_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "merchant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Merchant"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Days"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  },
                  "title": "Response List Transactions Me Transactions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Transactions",
        "tags": [
          "customers"
        ]
      }
    },
    "/me/transactions/{transaction_id}": {
      "get": {
        "operationId": "get_transaction_me_transactions__transaction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Transaction",
        "tags": [
          "customers"
        ]
      }
    },
...
    "/v1/disputes": {
      "get": {
        "description": "The customer's own cases, newest first.",
        "operationId": "list_disputes_v1_disputes_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(open|auto_resolved|escalated|in_progress)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Dispute"
                  },
                  "title": "Response List Disputes V1 Disputes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Disputes",
        "tags": [
          "disputes"
        ]
      },
      "post": {
        "description": "Open the case. One per (customer, transaction): reporting the same transaction again returns that case.",
        "operationId": "create_dispute_v1_disputes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
...
    "/v1/me": {
      "get": {
        "operationId": "get_me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/v1/me/transactions": {
      "get": {
        "description": "Newest first, a page at a time. Each row says whether it already has a dispute case.",
        "operationId": "list_transactions_v1_me_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "product_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "merchant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Merchant"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 300,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Transactions",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/me/transactions/{transaction_id}": {
      "get": {
        "operationId": "get_transaction_v1_me_transactions__transaction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Transaction",
        "tags": [
          "v1"
        ]
      }
    }
  }
}